requirements/local.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/local.in -o requirements/local.txt
across-server-openapi-python @ git+ssh://git@github.com/ACROSS-Team/across-server-openapi-python.git@d531116c9719854783fea15c5065213297ffbad0
    # via -r requirements/base.in
across-tools @ git+ssh://git@github.com/ACROSS-Team/across-tools.git@1bfa5cb0bdee2afd8a040f5f3defde0b7350471e
    # via -r requirements/base.in
alembic==1.14.0
    # via -r requirements/base.in
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   asyncer
    #   httpx
    #   starlette
    #   watchfiles
astropy==7.0.0
    # via
    #   -r requirements/base.in
    #   across-tools
    #   astropy-healpix
    #   astroquery
    #   healpy
    #   pyvo
astropy-healpix==1.0.3
    # via -r requirements/base.in
astropy-iers-data==0.2024.11.25.0.34.48
    # via astropy
astroquery==0.4.10
    # via
    #   -r requirements/base.in
    #   across-tools
asyncer==0.0.8
    # via -r requirements/base.in
asyncpg==0.30.0
    # via -r requirements/base.in
attrs==24.2.0
    # via rush
beautifulsoup4==4.13.4
    # via
    #   -r requirements/base.in
    #   astroquery
boto3==1.35.71
    # via
    #   -r requirements/base.in
    #   swifttools
botocore==1.35.71
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.38.46
    # via types-boto3
cachetools==5.5.0
    # via -r requirements/base.in
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
    #   skyfield
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
click-spinner==0.1.10
    # via fastapi-utilities
coverage==7.6.8
    # via pytest-cov
croniter==1.4.1
    # via fastapi-utilities
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via
    #   -r requirements/base.in
    #   fastapi
fastapi==0.112.0
    # via
    #   -r requirements/base.in
    #   -r requirements/local.in
    #   fastapi-utilities
    #   fastapi-utils
    #   pytest-fastapi-deps
fastapi-cli==0.0.5
    # via fastapi
fastapi-utilities==0.3.1
    # via -r requirements/base.in
fastapi-utils==0.8.0
    # via -r requirements/base.in
filelock==3.16.1
    # via virtualenv
geoalchemy2==0.15.2
    # via -r requirements/base.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
healpy==1.17.3
    # via
    #   -r requirements/base.in
    #   across-tools
html5lib==1.1
    # via astroquery
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via
    #   -r requirements/base.in
    #   across-tools
    #   fastapi
    #   spacetrack
identify==2.6.3
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
jaraco-classes==3.4.0
    # via keyring
jaraco-context==6.0.1
    # via keyring
jaraco-functools==4.1.0
    # via keyring
jinja2==3.1.4
    # via fastapi
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jplephem==2.22
    # via skyfield
keyring==25.6.0
    # via astroquery
logbook==1.8.0
    # via spacetrack
mako==1.3.6
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.7.0
    # via
    #   jaraco-classes
    #   jaraco-functools
mypy==1.13.0
    # via -r requirements/types.in
mypy-extensions==1.0.0
    # via
    #   mypy
    #   typing-inspect
nodeenv==1.9.1
    # via pre-commit
numpy==2.1.3
    # via
    #   across-tools
    #   astropy
    #   astropy-healpix
    #   astroquery
    #   healpy
    #   jplephem
    #   pandas
    #   pandas-stubs
    #   pyerfa
    #   shapely
    #   skyfield
    #   spiceypy
    #   swifttools
    #   types-shapely
packaging==24.2
    # via
    #   astropy
    #   geoalchemy2
    #   pytest
pandas==2.2.3
    # via
    #   -r requirements/base.in
    #   swifttools
pandas-stubs==2.2.3.241126
    # via -r requirements/base.in
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via -r requirements/lint.in
psutil==5.9.8
    # via fastapi-utils
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic==2.10.2
    # via
    #   across-server-openapi-python
    #   across-tools
    #   fastapi
    #   fastapi-utilities
    #   fastapi-utils
    #   pydantic-settings
pydantic-core==2.27.1
    # via pydantic
pydantic-settings==2.6.1
    # via
    #   -r requirements/base.in
    #   across-tools
pyerfa==2.0.1.5
    # via astropy
pygments==2.18.0
    # via rich
pyjwt==2.10.1
    # via -r requirements/base.in
pytest==8.3.3
    # via
    #   -r requirements/test.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-fastapi-deps
pytest-asyncio==0.25.1
    # via
    #   -r requirements/base.in
    #   -r requirements/test.in
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-fastapi-deps==0.2.3
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   across-server-openapi-python
    #   botocore
    #   croniter
    #   pandas
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via swifttools
python-multipart==0.0.19
    # via fastapi
pytz==2024.2
    # via pandas
pyvo==1.6.2
    # via astroquery
pyyaml==6.0.2
    # via
    #   astropy
    #   pre-commit
    #   uvicorn
represent==2.1
    # via spacetrack
requests==2.32.3
    # via
    #   astroquery
    #   pyvo
    #   swifttools
rich==13.9.4
    # via typer
rsa==4.9.1
    # via python-jose
ruff==0.8.1
    # via -r requirements/lint.in
rush==2021.4.0
    # via spacetrack
s3transfer==0.10.4
    # via boto3
sgp4==2.23
    # via
    #   -r requirements/base.in
    #   across-tools
    #   skyfield
shapely==2.0.6
    # via
    #   -r requirements/base.in
    #   across-tools
shellingham==1.5.4
    # via typer
six==1.16.0
    # via
    #   ecdsa
    #   html5lib
    #   python-dateutil
skyfield==1.49
    # via -r requirements/base.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   spacetrack
soupsieve==2.7
    # via beautifulsoup4
spacetrack==1.3.1
    # via
    #   -r requirements/base.in
    #   across-tools
spiceypy==6.0.2
    # via across-tools
sqlalchemy==2.0.36
    # via
    #   -r requirements/base.in
    #   alembic
    #   fastapi-utilities
    #   geoalchemy2
starlette==0.37.2
    # via fastapi
structlog==25.2.0
    # via -r requirements/base.in
swifttools==3.0.23
    # via -r requirements/base.in
tabulate==0.9.0
    # via swifttools
typer==0.14.0
    # via fastapi-cli
types-awscrt==0.27.6
    # via botocore-stubs
types-boto3==1.40.15
    # via -r requirements/types.in
types-boto3-ssm==1.40.0
    # via types-boto3
types-cachetools==5.5.0.20240820
    # via -r requirements/types.in
types-pyasn1==0.6.0.20240913
    # via types-python-jose
types-python-jose==3.3.4.20240106
    # via -r requirements/types.in
types-pytz==2025.1.0.20250204
    # via pandas-stubs
types-requests==2.32.0.20241016
    # via -r requirements/types.in
types-s3transfer==0.13.0
    # via types-boto3
types-shapely==2.0.0.20241112
    # via -r requirements/types.in
typing-extensions==4.12.2
    # via
    #   across-server-openapi-python
    #   alembic
    #   beautifulsoup4
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via -r requirements/base.in
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via
    #   across-server-openapi-python
    #   botocore
    #   requests
    #   types-requests
uvicorn==0.30.6
    # via
    #   -r requirements/base.in
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
virtualenv==20.28.0
    # via pre-commit
watchfiles==1.0.0
    # via uvicorn
webencodings==0.5.1
    # via html5lib
websockets==14.1
    # via uvicorn
